requirements.txt
pandas ~= 1.3.5
scikit-learn~=1.0.2
PyYAML~=6.0
pyreadr~=0.4.4
ipywidgets~=7.6.5
tqdm
modin[dask]~=0.12.1
rpy2
matplotlib~=3.5.1
nmf-torch
halo
numpy==1.21
scipy~=1.7.3
jupyterlab
plotly~=5.6.0
scikit-optimize
kaleido
umap-learn
more-itertools
ipython~=8.0.1
imbalanced-learn